Handled updating/deleting already deleted items
Handled the case when user tries to update or delete locally cached item that is already deleted on the server.
The case is handled by informing the user with an appropriate message, so the user knows that his data is not sync with the server, and need to refresh his lists.